requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-import-export==3.3.1
django-jazzmin==2.6.0
django-shortuuidfield==0.1.3
et-xmlfile==1.1.0
gunicorn==21.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.9
PyYAML==6.0.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
